// src/app/core/models/reports.ts
export class reportNomina {
    businessName: string;
    branch: string;
    job: string;
    employerRegistration: string;
    names: string;
    surnames: string;
    RFC: string;
    employeeNumber: string;
    medicalNumber: string;
    department: string;
    salary: number;
    laboryDays: number;
    period: string;
    period2: string;

    constructor(){
        this.businessName = '';
        this.branch = '';
        this.job = '';
        this.employerRegistration = '';
        this.names = '';
        this.surnames = '';
        this.RFC = '';
        this.employeeNumber = '';
        this.medicalNumber = '';
        this.department = '';
        this.salary = 0;
        this.laboryDays = 0;
        this.period = '';
        this.period2 = '';
    }
}